Separate column tooltip from core table cell
In preparation for adding more functionality to the column tooltip HOC,
separate it from TableCell. No changed functionality.
